move utils and improve exclude
